re-organize printer code into separate package
