+++ print/json.go
@@ -1,0 +1,18 @@
+package print
+
+import (
+	"encoding/json"
+
+	"github.com/mlange-42/dirstat/tree"
+)
+
+// JSONPrinter prints a tree in JSON format
+type JSONPrinter[T any] struct{}
+
+func (p JSONPrinter[T]) Print(t *tree.Tree[T]) string {
+	tt, err := json.MarshalIndent(t, "", "    ")
+	if err != nil {
+		panic(err)
+	}
+	return string(tt[:])
+}
